Restructure tests back to per-topic files
The previous sweep collapsed every Express-dependent topic into a single
test/lib/shared.ts, losing the per-file granularity that the original
main/test/*.ts layout offered. This commit decomposes shared.ts into
topic-specific modules under test/lib/ (compressed, css, html, large-head,
xml), each exporting a single run<Topic>Tests(express) function. The two
entry points (express4.test.ts / express5.test.ts) import all five and run
them as describe blocks. test/match.test.ts is unchanged (Express-independent).

The shared HTDOCS path, assert_match helper and ExpressFactory type now
live in test/lib/util.ts. Type-only imports use `import type` so that
Node's strip-only TypeScript loader keeps them out of runtime resolution.

Also runs fixpack against package.json (keywords reordered alphabetically).
